compile lambda to index.js (worked on aws)
