+++ .gitignore
@@ -24,3 +24,4 @@ npm-debug.log*
 /src/lambda/package-lock.json
 /src/lambda/yarn.lock
 /src/lambda/build
+/src/lambda/index.js
