+++ src/lambda/package.json
@@ -2,7 +2,7 @@
   "name": "radius-tracker-lambda-worker",
   "version": "0.0.1",
   "description": "Lambda worker that searching for snowflakes",
-  "main": "build/worker.js",
+  "main": "index.js",
   "scripts": {
     "compile": "tsc -p ./tsconfig.json"
   },
@@ -13,11 +13,17 @@
   "license": "MIT",
   "devDependencies": {
     "@types/aws-lambda": "^8.10.93",
-    "@types/node": "^17.0.23",
     "aws-sdk": "^2.1103.0",
     "typescript": "^4.6.3"
   },
   "dependencies": {
-    "radius-tracker": "^0.0.12"
+    "@octokit/rest": "^18.12.0",
+    "@ts-morph/common": "^0.13.0",
+    "buffer": "^6.0.3",
+    "isomorphic-git": "^1.12.1",
+    "memfs": "^3.4.1",
+    "radius-tracker": "^0.0.12",
+    "ts-morph": "^13.0.2",
+    "tslib": "^2.3.1"
   }
 }
